Widen TallyTask::filter_idx to int64_t
TallyTask stores the flat tally results_ index for each source-region/energy-group element. With more than 2^31 filter bins the int member truncated the now-64-bit filter index taken from FilterBinIter::index_, sending random ray scores to the wrong (or out-of-bounds) results_ bins. Widen the member and the constructor parameter to int64_t.

+++ include/openmc/random_ray/source_region.h
@@ -51,10 +51,10 @@ inline void hash_combine(size_t& seed, const size_t v)
 // every iteration.
 struct TallyTask {
   int tally_idx;
-  int filter_idx;
+  int64_t filter_idx;
   int score_idx;
   int score_type;
-  TallyTask(int tally_idx, int filter_idx, int score_idx, int score_type)
+  TallyTask(int tally_idx, int64_t filter_idx, int score_idx, int score_type)
     : tally_idx(tally_idx), filter_idx(filter_idx), score_idx(score_idx),
       score_type(score_type)
   {}
